round the max memory out

+++ examples/IBFE/explicit/ex4/example.cpp
@@ -68,12 +68,8 @@
 // Set up application namespace declarations
 #include <ibamr/app_namespaces.h>
 
-
-
 #include <sys/time.h>
 #include <sys/resource.h>
-
-
 
 // Elasticity model data.
 namespace ModelData
@@ -477,7 +473,7 @@ run_example(int argc, char** argv)
             rusage r_usage;
             getrusage(RUSAGE_SELF, &r_usage);
             long mem_usage = r_usage.ru_maxrss;
-            pout << "Current memory max:" << SAMRAI_MPI::sumReduction(double(mem_usage)) << "\n";
+            pout << "Current memory max:" << long(SAMRAI_MPI::sumReduction(double(mem_usage))) << "\n";
 
             pout << "\n";
             pout << "+++++++++++++++++++++++++++++++++++++++++++++++++++\n";
